backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.0
anthropic==0.34.0
supabase==2.7.0
httpx==0.27.0
python-multipart==0.0.9
pillow==10.4.0
pydantic==2.8.0
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
apscheduler==3.10.4
websockets==13.0
geopy==2.4.1
